+++ src/gitlab.ts
@@ -1,4 +1,4 @@
-import { CONFIG } from "./config";
+import { CONFIG } from './config';
 import {
   Branch,
   Issue,
@@ -6,13 +6,14 @@ import {
   MergeRequest,
   Project,
   User,
-} from "./models/gitlab.models";
-import { Commit } from "./models/gitlab/commit.models";
-import { Job } from "./models/gitlab/job.models";
-import { FullMergeRequest } from "./models/gitlab/merge-request.models";
-import { callApiFactory, dashify } from "./utils";
+} from './models/gitlab.models';
+import { Commit } from './models/gitlab/commit.models';
+import { Job } from './models/gitlab/job.models';
+import { FullMergeRequest } from './models/gitlab/merge-request.models';
+import { Pipeline } from './models/gitlab/pipeline.models';
+import { callApiFactory, dashify } from './utils';
 
-const callApi = callApiFactory("GitLab");
+const callApi = callApiFactory('GitLab');
 
 export function getGitLabBranchNameFromIssueNumberAndTitleAndTaskId(
   issueNumber: number,
@@ -26,7 +27,7 @@ export class GitLab {
   constructor(public projectId: string) {}
 
   public getProject() {
-    return callApi<Project>("get", `/projects/${this.projectId}`);
+    return callApi<Project>('get', `/projects/${this.projectId}`);
   }
 
   public async getDefaultBranchName() {
@@ -36,82 +37,75 @@ export class GitLab {
 
   public getIssue(issueNumber: string) {
     return callApi<Issue>(
-      "get",
+      'get',
       `/projects/${this.projectId}/issues/${issueNumber}`
     );
   }
 
   public getMergeRequest(mergeRequestNumber: string | number) {
     return callApi<FullMergeRequest>(
-      "get",
+      'get',
       `/projects/${this.projectId}/merge_requests/${mergeRequestNumber}`
     );
   }
 
   public getCommit(sha: string) {
     return callApi<Commit>(
-      "get",
+      'get',
       `/projects/${this.projectId}/repository/commits/${sha}`
     );
   }
 
   public getEndingAssignee() {
     if (!CONFIG.EndingAssignee) {
-      throw Error("No ending assignee was set");
+      throw Error('No ending assignee was set');
     }
     return callApi<User[]>(
-      "get",
+      'get',
       `/users?username=${CONFIG.EndingAssignee}`
     ).then((users) => users[0]);
   }
 
   public listProjectLabels() {
     return callApi<Label[]>(
-      "get",
+      'get',
       `/projects/${this.projectId}/labels?per_page=100`
     );
   }
 
   public listMergeRequestsWillCloseIssueOnMerge(issueNumber: string) {
     return callApi<MergeRequest[]>(
-      "get",
+      'get',
       `/projects/${this.projectId}/issues/${issueNumber}/closed_by`
     );
   }
 
   public listPipelineJobs(pipelineId: number) {
     return callApi<Job[]>(
-      "get",
+      'get',
       `/projects/${this.projectId}/pipelines/${pipelineId}/jobs`
-    );
-  }
-
-  public getPipeline(pipelineId: string) {
-    return callApi<any>(
-      "get",
-      `/projects/${this.projectId}/pipelines/${pipelineId}`
     );
   }
 
   public listPipelines(sha: string, ref: string) {
     return callApi<Pipeline[]>(
-      "get",
+      'get',
       `/projects/${this.projectId}/pipelines/?sha=${sha}&ref=${ref}`
     );
   }
 
   public async createIssue(title: string, description: string, labels: any[]) {
-    return callApi<Issue>("post", `/projects/${this.projectId}/issues`, {
+    return callApi<Issue>('post', `/projects/${this.projectId}/issues`, {
       title: title,
       description: description,
       assignee_ids: await this.getUserId(),
-      labels: labels.join(","),
+      labels: labels.join(','),
     });
   }
 
   public async createBranch(branch: string) {
     return callApi<Branch>(
-      "post",
+      'post',
       `/projects/${this.projectId}/repository/branches`,
       {
         branch,
@@ -127,14 +121,14 @@ export class GitLab {
     labels: any[]
   ) {
     return callApi<MergeRequest>(
-      "post",
+      'post',
       `/projects/${this.projectId}/merge_requests`,
       {
         source_branch: branch,
         target_branch: await this.getDefaultBranchName(),
         title: `Draft: Resolve "${issueTitle}"`,
         description: `Close #${issueNumber}`,
-        labels: labels.join(","),
+        labels: labels.join(','),
       }
     );
   }
@@ -144,17 +138,17 @@ export class GitLab {
   ) {
     const assignee = await this.getEndingAssignee();
     await callApi(
-      "put",
+      'put',
       `/projects/${this.projectId}/merge_requests/${merge_request.iid}`,
       {
-        title: merge_request.title.replace("WIP: ", "").replace("Draft: ", ""),
+        title: merge_request.title.replace('WIP: ', '').replace('Draft: ', ''),
         assignee_id: assignee.id,
       }
     );
   }
 
   private async getUserId() {
-    const user = await callApi<User>("get", "/user");
+    const user = await callApi<User>('get', '/user');
     return user.id;
   }
 }