+++ src/models/gitlab.models.ts
@@ -2,5 +2,6 @@ export { Branch } from './gitlab/branch.models';
 export { Issue } from './gitlab/issue.models';
 export { Label } from './gitlab/label.models';
 export { MergeRequest } from './gitlab/merge-request.models';
+export { Pipeline } from './gitlab/pipeline.models';
 export { Project } from './gitlab/project.models';
 export { User } from './gitlab/user.models';